크루생성 api 연결 완료

+++ wupitch/wupitch/Network/APIModels/Crew/MakeCrewRequest.swift
@@ -33,5 +33,5 @@ struct MakeCrewRequest: Codable {
 struct ScheduleList: Codable {
     let dayIdx: Int
     let startTime: Double
-    let endTime: Int
+    let endTime: Double
 }
